PABR Update

+++ src/leansdr/framework.h
@@ -28,24 +28,24 @@
 #endif
 
 namespace leansdr {
-  
+
   void fatal(const char *s) { perror(s); exit(1); }
   void fail(const char *s) { fprintf(stderr, "** %s\n", s); exit(1); }
-  
+
   //////////////////////////////////////////////////////////////////////
   // DSP framework
   //////////////////////////////////////////////////////////////////////
-  
+
   // [pipebuf] is a FIFO buffer with multiple readers.
   // [pipewriter] is a client-side hook for writing into a [pipebuf].
   // [pipereader] is a client-side hook reading from a [pipebuf].
   // [runnable] is anything that moves data between [pipebufs].
   // [scheduler] is a global context which invokes [runnables] until fixpoint.
-  
+
   static const int MAX_PIPES = 64;
   static const int MAX_RUNNABLES = 64;
   static const int MAX_READERS = 8;
-  
+
   struct pipebuf_common {
     virtual int sizeofT() { return 0; }
     virtual long long hash() { return 0; }
@@ -63,7 +63,7 @@ namespace leansdr {
     ~runnable_common() { fprintf(stderr, "Deallocating %s !\n", name); }
 #endif
   };
-  
+
   struct window_placement {
     const char *name; // NULL to terminate
     int x, y, w, h;
@@ -111,7 +111,7 @@ namespace leansdr {
       for ( int i=0; i<npipes; ++i ) h += (1+i)*pipes[i]->hash();
       return h;
     }
-    
+
     void dump() {
       fprintf(stderr, "\n");
       size_t total_bufs = 0;
@@ -120,7 +120,7 @@ namespace leansdr {
 	      (unsigned long)total_bufs/1024);
     }
   };
-  
+
   struct runnable : runnable_common {
     runnable(scheduler *_sch, const char *name)
       : runnable_common(name), sch(_sch) {
@@ -129,7 +129,7 @@ namespace leansdr {
   protected:
     scheduler *sch;
   };
-  
+
   template<typename T>
   struct pipebuf : pipebuf_common {
     T *buf;
@@ -161,13 +161,13 @@ namespace leansdr {
       return total_written + total_read;
     }
     void dump(size_t *total_bufs) {
-      if ( total_written < 10000 ) 
+      if ( total_written < 10000 )
 	fprintf(stderr, ".%-16s : %4ld/%4ld", name,
 		total_read, total_written);
-      else if ( total_written < 1000000 ) 
+      else if ( total_written < 1000000 )
 	fprintf(stderr, ".%-16s : %3ldk/%3ldk", name,
 		total_read/1000, total_written/1000);
-      else 
+      else
 	fprintf(stderr, ".%-16s : %3ldM/%3ldM", name,
 		total_read/1000000, total_written/1000000);
       *total_bufs += (end-buf) * sizeof(T);
@@ -186,7 +186,7 @@ namespace leansdr {
     ~pipebuf() { fprintf(stderr, "Deallocating %s !\n", name); }
 #endif
   };
-  
+
   template<typename T>
   struct pipewriter {
     pipebuf<T> &buf;
@@ -195,7 +195,7 @@ namespace leansdr {
       if ( min_write > buf.min_write ) buf.min_write = min_write;
     }
     // Return number of items writable at this->wr, 0 if full.
-    unsigned long writable() {
+    size_t writable() {
       if ( buf.end-buf.wr < buf.min_write ) buf.pack();
       return buf.end - buf.wr;
     }
@@ -236,7 +236,7 @@ namespace leansdr {
     pipebuf<T> &buf;
     int id;
     pipereader(pipebuf<T> &_buf) : buf(_buf), id(_buf.add_reader()) { }
-    unsigned long readable() { return buf.wr - buf.rds[id]; }
+    size_t readable() { return buf.wr - buf.rds[id]; }
     T *rd() { return buf.rds[id]; }
     void read(unsigned long n) {
       if ( buf.rds[id]+n > buf.wr ) {
@@ -247,10 +247,10 @@ namespace leansdr {
       buf.total_read += n;
     }
   };
-  
+
   // Math functions for templates
-  
-  template<typename T> T gen_sqrt(T x);  
+
+  template<typename T> T gen_sqrt(T x);
   inline float           gen_sqrt(float x)        { return sqrtf(x); }
   inline unsigned int    gen_sqrt(unsigned int x) { return sqrtl(x); }
   inline long double     gen_sqrt(long double x)  { return sqrtl(x); }
@@ -260,12 +260,12 @@ namespace leansdr {
   inline int             gen_abs(int x)      { return abs(x); }
   inline long int        gen_abs(long int x) { return labs(x); }
 
-  template<typename T> T gen_hypot(T x, T y);  
+  template<typename T> T gen_hypot(T x, T y);
   inline float           gen_hypot(float x, float y) { return hypotf(x,y); }
   inline long double     gen_hypot(long double x, long double y)
   { return hypotl(x,y); }
 
-  template<typename T> T gen_atan2(T y, T x);  
+  template<typename T> T gen_atan2(T y, T x);
   inline float           gen_atan2(float y, float x) { return atan2f(y,x); }
   inline long double     gen_atan2(long double y, long double x)
   { return atan2l(y,x); }
